rename packages. Code that index files to JDBC / ES / BQ is all now located in index package. Classes are consistently named [ES/JDBC/BQ]Load[Config/Job]
